+++ src/ConfigurationContext.tsx
@@ -14,6 +14,7 @@ export interface Configuration {
   displaySymptomHistory: boolean
   displaySelfAssessment: boolean
   displayAgeVerification: boolean
+  enableProductAnalytics: boolean
   emergencyPhoneNumber: string
   findATestCenterUrl: string | null
   healthAuthorityAdviceUrl: string
@@ -23,7 +24,6 @@ export interface Configuration {
   healthAuthorityLegalPrivacyPolicyUrl: string | null
   healthAuthorityName: string
   healthAuthorityPrivacyPolicyUrl: string
-  healthAuthoritySupportsAnalytics: boolean
   measurementSystem: MeasurementSystem
   minimumAge: string
   regionCodes: string[]
@@ -41,6 +41,7 @@ const initialState: Configuration = {
   displaySelfAssessment: false,
   displayAgeVerification: false,
   emergencyPhoneNumber: "",
+  enableProductAnalytics: false,
   findATestCenterUrl: null,
   healthAuthorityAdviceUrl: "",
   healthAuthorityCovidDataUrl: null,
@@ -49,7 +50,6 @@ const initialState: Configuration = {
   healthAuthorityLegalPrivacyPolicyUrl: "",
   healthAuthorityName: "",
   healthAuthorityPrivacyPolicyUrl: "",
-  healthAuthoritySupportsAnalytics: false,
   measurementSystem: "Imperial" as const,
   minimumAge: "18",
   regionCodes: [],
@@ -80,13 +80,12 @@ const ConfigurationProvider: FunctionComponent = ({ children }) => {
   const displaySymptomHistory = env.DISPLAY_SYMPTOM_HISTORY === "true"
   const displaySelfAssessment = env.DISPLAY_SELF_ASSESSMENT === "true"
   const displayAgeVerification = env.DISPLAY_AGE_VERIFICATION === "true"
+  const enableProductAnalytics = env.ENABLE_PRODUCT_ANALYTICS === "true"
 
   const measurementSystem =
     env.MEASUREMENT_SYSTEM === "metric" ? "Metric" : "Imperial"
 
   const minimumAge = env.MINIMUM_AGE
-
-  const healthAuthoritySupportsAnalytics = Boolean(env.MATOMO_URL)
 
   const appDownloadLink = env.SHARE_APP_LINK
   const appPackageName = Platform.select({
@@ -110,6 +109,7 @@ const ConfigurationProvider: FunctionComponent = ({ children }) => {
         displaySymptomHistory,
         displaySelfAssessment,
         displayAgeVerification,
+        enableProductAnalytics,
         emergencyPhoneNumber,
         findATestCenterUrl,
         healthAuthorityAdviceUrl,
@@ -119,7 +119,6 @@ const ConfigurationProvider: FunctionComponent = ({ children }) => {
         healthAuthorityLegalPrivacyPolicyUrl: legalPrivacyPolicyUrl || null,
         healthAuthorityName,
         healthAuthorityPrivacyPolicyUrl,
-        healthAuthoritySupportsAnalytics,
         measurementSystem,
         minimumAge,
         regionCodes,
