+++ src/ExposureContext.tsx
@@ -12,7 +12,6 @@ import { failureResponse, OperationResponse } from "./OperationResponse"
 import { ExposureKey } from "./exposureKey"
 import { ExposureInfo } from "./exposure"
 import { checkForNewExposures as detectExposures } from "./gaen/nativeModule"
-import { useProductAnalyticsContext } from "./ProductAnalytics/Context"
 
 type Posix = number
 const { exposureEventsStrategy } = gaenStrategy
@@ -62,7 +61,6 @@ const ExposureProvider: FunctionComponent = ({ children }) => {
     exposureInfoSubscription,
     getLastDetectionDate,
   } = exposureEventsStrategy
-  const { trackEvent } = useProductAnalyticsContext()
 
   const [exposureInfo, setExposureInfo] = useState<ExposureInfo>([])
 
@@ -88,12 +86,10 @@ const ExposureProvider: FunctionComponent = ({ children }) => {
   useEffect(() => {
     const subscription = exposureInfoSubscription(
       (exposureInfo: ExposureInfo) => {
-        trackEvent("epi_analytics", "en_notification_received")
         setExposureInfo(exposureInfo)
         getLastExposureDetectionDate()
       },
     )
-
     getLastExposureDetectionDate()
 
     return subscription.remove
