+++ src/Settings/index.spec.tsx
@@ -28,7 +28,7 @@ describe("Settings", () => {
 
       const { getByLabelText } = render(<SettingsScreen />)
 
-      const deleteMyDataButton = getByLabelText("Delete My Data")
+      const deleteMyDataButton = getByLabelText("Delete my data")
       fireEvent.press(deleteMyDataButton)
       expect(navigateSpy).toHaveBeenCalledWith(
         SettingsStackScreens.DeleteConfirmation,
