+++ src/factories/configurationContext.ts
@@ -11,6 +11,7 @@ export default Factory.define<Configuration>(() => ({
   displaySymptomHistory: false,
   displaySelfAssessment: false,
   displayAgeVerification: false,
+  enableProductAnalytics: false,
   emergencyPhoneNumber: "emergencyPhoneNumber",
   findATestCenterUrl: "findATestCenterUrl",
   healthAuthorityAdviceUrl: "authorityAdviceUrl",
@@ -20,7 +21,6 @@ export default Factory.define<Configuration>(() => ({
   healthAuthorityName: "authorityName",
   healthAuthorityPrivacyPolicyUrl: "authorityPrivacyPolicyUrl",
   healthAuthorityLegalPrivacyPolicyUrl: "authorityLegalPrivacyPolicyUrl",
-  healthAuthoritySupportsAnalytics: false,
   measurementSystem: "Imperial",
   minimumAge: "18",
   regionCodes: ["REGION"],
