+++ src/locales/en.json
@@ -391,25 +391,30 @@
       "weakened_immune_system": "Weakened immune system or taking medications that may cause immune suppression"
     }
   },
+  "product_analytics": {
+    "not_sharing_anonymized_data": "Not sharing anonymized data",
+    "share_anonymized_data": "Share Anonymized Data",
+    "share_anonymized_data_disclaimer": "By proceeding, you agree to share anonymized answers to screening questions with the {{healthAuthorityName}}",
+    "share_data_para_1": "Anonymous and aggregated data about how communities are using this app helps the {{healthAuthorityName}} Commission, analyze, understand, and contain the spread of COVID-19.",
+    "share_data_para_2": "This data contains no personally-identifying information.",
+    "sharing_anonymized_data": "Sharing anonymized data",
+    "privacy_policy": "Privacy Policy",
+    "stop_sharing_data": "Stop Sharing Data",
+    "you_are_sharing_data": "You are sharing anonymized data",
+    "i_understand_and_consent": "I Understand and Consent"
+  },
   "settings": {
     "data_deleted": "Data deleted",
     "delete_data_disclosure1": "When you tap the \"Delete My Data\" button below, the app will delete the Symptom History data that is stored on your phone.",
     "delete_data_disclosure2": "This does not delete the data created or collected by your phone for Exposure Notifications.\n\nYou can delete this additional data in your device Settings.",
     "delete_data_note": "Note",
-    "delete_my_data": "Delete My Data",
+    "delete_my_data": "Delete my data",
     "errors": {
       "deleting_data": "Sorry, we could not delete your data"
     },
-    "here": "here",
-    "i_understand_and_consent": "I Understand and Consent",
-    "not_sharing_anonymized_data": "Not sharing anonymized data",
-    "please_view_privacy_policy": "Please view their privacy policy",
-    "share_anonymized_data": "Share Anonymized Data",
-    "share_anonymized_data_disclaimer": "By proceeding, you agree to share anonymized answers to screening questions with the {{healthAuthorityName}}",
-    "share_anonymized_data_para": "Your anonymous data helps the {{healthAuthorityName}} Commission analyze, track, and contain the spread of COVID-19. Your data will be aggregated with data from other users who choose to share.\n\nYour data contains your survey responses and zipcode, but no additional or personally-identifying information.\n\n The {{healthAuthorityName}} will retain your anonymous data for a specific period of time.",
+    "share_anonymized_data": "Share anonymized data",
     "sharing_anonymized_data": "Sharing anonymized data",
-    "stop_sharing_data": "Stop Sharing Data",
-    "you_are_sharing_data": "You are sharing anonymized data"
+    "not_sharing_anonymized_data": "Not sharing anonymized data"
   },
   "symptom_history": {
     "did_not_feel_well": "You did not feel well",
